# gmmsales-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GMMSales API Documentation",
        "description": "Dokumentasi API untuk aplikasi GMMSales - sistem pencatatan kunjungan customer oleh sales lapangan dengan dokumentasi GPS dan foto.",
        "contact": {
            "name": "Rakha Hafizh",
            "email": "rakhahafizh@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "GMMSales Local Development Server"
        }
    ],
    "paths": {
...
        "/api/admin/sales": {
            "get": {
                "tags": [
                    "Admin - Sales"
                ],
                "summary": "Daftar semua sales (Admin)",
                "operationId": "4b54ffd4594e487bb4e0fb53e89c17e1",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "wilayah_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar sales berhasil diambil"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Bukan role admin"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin - Sales"
                ],
                "summary": "Tambah sales baru (Admin)",
                "operationId": "0543d6fd9b676958243637408b875f97",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "username",
                                    "password",
                                    "nomor_telepon",
                                    "wilayah_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Budi Santoso"
                                    },
                                    "username": {
                                        "type": "string",
                                        "example": "budi"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "nomor_telepon": {
                                        "type": "string",
                                        "example": "081234567890"
                                    },
                                    "wilayah_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sales berhasil dibuat"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Bukan role admin"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/sales/{id}": {
            "get": {
                "tags": [
                    "Admin - Sales"
                ],
                "summary": "Detail sales (Admin)",
                "operationId": "8dc6fc9fa57d07e47233e0ec0100ac53",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail sales berhasil diambil"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Bukan role admin"
                    },
                    "404": {
                        "description": "Sales tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin - Sales"
                ],
                "summary": "Edit data sales (Admin)",
                "description": "Admin mengubah data sales. Field is_active tidak dapat diubah secara manual karena bersifat auto-managed (set true saat sales submit customer, reset false tiap tengah malam).",
                "operationId": "3e1bcc596776b04ff273d587a1ebf187",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "nomor_telepon": {
                                        "type": "string"
                                    },
                                    "wilayah_id": {
                                        "type": "integer"
                                    },
                                    "password": {
                                        "description": "Opsional, untuk reset password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sales berhasil diupdate"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Bukan role admin"
                    },
                    "404": {
                        "description": "Sales tidak ditemukan"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin - Sales"
                ],
                "summary": "Hapus sales (soft delete)",
                "operationId": "6110e9f9991475454325e267cd3f509d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sales berhasil dihapus"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Bukan role admin"
                    },
                    "404": {
                        "description": "Sales tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token Sanctum tanpa prefix 'Bearer ' (cukup tokennya saja)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}